+++ src/main/java/com/kuna_backend/dtos/checkout/CheckoutItemDto.java
@@ -5,13 +5,13 @@ public class CheckoutItemDto {
     private String productName;
     private int quantity;
     private double price;
-    private int productVariationId;
-    private int clientId;
+    private Long productVariationId;
+    private Long clientId;
 
     public CheckoutItemDto() {
     }
 
-    public CheckoutItemDto(String productName, int quantity, double price, int productVariationId, int clientId) {
+    public CheckoutItemDto(String productName, int quantity, double price, Long productVariationId, Long clientId) {
         this.productName = productName;
         this.quantity = quantity;
         this.price = price;
@@ -43,19 +43,19 @@ public class CheckoutItemDto {
         this.price = price;
     }
 
-    public int getProductVariationId() {
+    public Long getProductVariationId() {
         return productVariationId;
     }
 
-    public void setProductVariationId(int productVariationId) {
+    public void setProductVariationId(Long productVariationId) {
         this.productVariationId = productVariationId;
     }
 
-    public int getClientId() {
+    public Long getClientId() {
         return clientId;
     }
 
-    public void setClientId(int clientId) {
+    public void setClientId(Long clientId) {
         this.clientId = clientId;
     }
 }
